+++ lib/navigation/navigation.dart
@@ -1,0 +1,20 @@
+import 'package:flutter/material.dart';
+import 'package:utranslator/builders/build_header.dart';
+import 'package:utranslator/builders/build_menu_items.dart';
+
+class Navigation extends StatelessWidget {
+  const Navigation({super.key});
+
+  @override
+  Widget build(BuildContext context) => Drawer(
+        child: SingleChildScrollView(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.stretch,
+            children: <Widget>[
+              buildHeader(context),
+              buildMenuItems(context),
+            ],
+          ),
+        ),
+      );
+}
